Work towards being able to use Sphynx from a DLL.  Need to figure out how to export STL object types, and I need to tag a lot of free functions with CAT_EXPORT.

+++ src/sphynx/ConnexionMap.cpp
@@ -292,7 +292,7 @@ void ConnexionMap::ShutdownAll()
 	lock.Release();
 
 	// For each Connexion object to release,
-	for (u32 ii = 0, size = connexions.size(); ii < size; ++ii)
+	for (u32 ii = 0, size = (u32)connexions.size(); ii < size; ++ii)
 	{
 		Connexion *conn = connexions[ii];
